requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
cryptography
psycopg2-binary